ss: fixes

+++ src/common/ss-client-logproxy.c
@@ -173,7 +173,7 @@ saicom_lp_state(void *userobj, void *sh, lws_ss_constate_t state,
 {
 	said_logproxy_t *lp = (said_logproxy_t *)userobj;
 
-	lwsl_info("%s: %s, ord 0x%x\n", __func__, lws_ss_state_name((int)state),
+	lwsl_info("%s: %s, ord 0x%x\n", __func__, lws_ss_state_name(state),
 		  (unsigned int)ack);
 
 	switch (state) {
